requirements.txt
numpy
torch==2.9.1
torchvision==0.24.1
pytorch-msssim==1.0.0
pytorchvideo==0.1.5
gradio==6.2.0
gradio_client==2.0.2
opencv-python==4.6.0.66
pillow==10.4.0
pillow_heif==0.15.0
matplotlib==3.10.8
matplotlib-inline==0.1.6
tqdm==4.65.0
moviepy==1.0.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.11.4
random-fourier-features-pytorch